Move associations to new package
